Move Czech app page components to server folder
