+++ apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/[first]/navod/[step]/page.tsx
@@ -1,6 +1,6 @@
 import { notFound } from "next/navigation";
 
-import { GuidePage } from "../../../../../../components/client/pages";
+import { GuidePage } from "../../../../../../components/server";
 
 export default async function Page({ params }: { params: Promise<{ step: string }> }) {
   const stepInt = Number.parseInt((await params).step);
