+++ apps/www.volebnikalkulacka.cz/app/(web)/(kalkulacka)/[first]/vysledek/page.tsx
@@ -1,4 +1,4 @@
-import { ResultPage } from "../../../../../components/client/pages";
+import { ResultPage } from "../../../../../components/server";
 
 export default function Page() {
   return <ResultPage />;
