+++ apps/www.volebnikalkulacka.cz/components/client/index.ts
@@ -1,5 +1,4 @@
 "use client";
 
 export * from "./embed-provider";
-export * from "./pages";
 export * from "./theme-provider";
