+++ apps/www.volebnikalkulacka.cz/components/server/pages/calculator/index.ts
