+++ apps/www.volebnikalkulacka.cz/components/server/pages/calculator/result.tsx
